refactor: extract suggestion machinery to its own hook
